refactor(ui): split EquipmentTab into concern partials (mechanical, no behavior change)
2220 lines in one file -> six partials by concern, following the
ImageRendererBase / LiveSessionTab convention: core (state/input/entry
views/bottom bar/shared helpers), ProfilePanel (section walk + section
renderers), DeviceList (scroll list + confirm strips + On|Off segment),
Telemetry (cooler pane + sparkline + mount status), DeviceSettings
(descriptor rows + Advanced sub-section), FilterTable (editor). Code is
moved verbatim; the class declaration gains `partial`. The stranded
RenderDeviceSettingsIfAny doc comment moved back onto its method.

Prepares the L2 layout-driven conversion (docs/plans/layout-driven-ui.md)
so the one-tree profile panel lands as a reviewable diff instead of a
rewrite buried in a 2k-line file.
