+++ src/TianWen.UI.Abstractions/EquipmentTab.cs
@@ -13,7 +13,7 @@ namespace TianWen.UI.Abstractions
     /// Renderer-agnostic Equipment / Profile tab. Handles profile summary,
     /// discovered devices, device assignment, filter editing, and OTA properties.
     /// </summary>
-    public class EquipmentTab<TSurface>(Renderer<TSurface> renderer) : PixelWidgetBase<TSurface>(renderer)
+    public partial class EquipmentTab<TSurface>(Renderer<TSurface> renderer) : PixelWidgetBase<TSurface>(renderer)
     {
         private readonly EquipmentContent _content = new EquipmentContent();
 
@@ -369,684 +369,6 @@ namespace TianWen.UI.Abstractions
             RenderBottomBar(appState, bottomBarRect, dpiScale, fontPath);
         }
 
-        // -----------------------------------------------------------------------
-        // Left panel: profile summary
-        // -----------------------------------------------------------------------
-
-        private void RenderProfilePanel(
-            GuiAppState appState,
-            RectF32 rect,
-            float dpiScale, string fontPath,
-            string? emojiFontPath = null,
-            LiveSessionState? liveSessionState = null)
-        {
-            var fontSize   = BaseFontSize * dpiScale;
-            var padding    = BasePadding * dpiScale;
-            var itemH      = BaseItemHeight * dpiScale;
-            var arrowW     = BaseArrowWidth * dpiScale;
-            var headerH    = BaseHeaderHeight * dpiScale;
-            var buttonH    = BaseButtonHeight * dpiScale;
-
-            var x = rect.X;
-            var y = rect.Y;
-            var w = rect.Width;
-            var h = rect.Height;
-
-            var profile = appState.ActiveProfile!;
-            var data = profile.Data;
-
-            var cursor = y + padding;
-
-            if (data is { } pd)
-            {
-                // The panel's vertical flow is a data-driven section list (TODO.md:57): the order lives in
-                // EquipmentContent.GetProfilePanelSections, not in a hardcoded cursor walk. Each section is
-                // dispatched to its renderer (which advances the cursor); runtime-conditional sections
-                // (mount/camera telemetry, device settings) self-gate to a no-op when their device isn't
-                // hub-connected / declares no settings, exactly as the inline calls did before.
-                foreach (var section in _content.GetProfilePanelSections(pd))
-                {
-                    cursor = RenderSection(section, appState, pd, profile, liveSessionState,
-                        x, cursor, w, y, h, dpiScale, fontPath, emojiFontPath,
-                        fontSize, padding, itemH, arrowW, headerH, buttonH);
-                }
-            }
-            else
-            {
-                cursor = RenderProfileHeaderSection(profile, x, cursor, w, fontPath, fontSize, padding, headerH);
-                FillRect(x + padding, cursor, w - padding * 2f, 1f, SeparatorColor);
-                cursor += padding;
-                DrawText(
-                    "Profile has no data.".AsSpan(),
-                    fontPath,
-                    x + padding, cursor, w - padding * 2f, itemH,
-                    fontSize, DimText, TextAlign.Near, TextAlign.Center);
-            }
-        }
-
-        /// <summary>
-        /// Dispatches one <see cref="PanelSection"/> to its renderer at the current cursor Y, returning the
-        /// new cursor. Static rows (slots / header / separator) go through the engine-backed helpers;
-        /// interactive + variable-height widgets stay as imperative section helpers. This is the single
-        /// switch the data-driven section loop (<see cref="EquipmentContent.GetProfilePanelSections"/>) walks,
-        /// replacing the old hardcoded cursor sequence.
-        /// </summary>
-        private float RenderSection(PanelSection section, GuiAppState appState, ProfileData pd, Profile profile,
-            LiveSessionState? liveSessionState, float x, float cursor, float w, float y, float h,
-            float dpiScale, string fontPath, string? emojiFontPath,
-            float fontSize, float padding, float itemH, float arrowW, float headerH, float buttonH)
-        {
-            switch (section)
-            {
-                case PanelSection.ProfileHeader:
-                    return RenderProfileHeaderSection(profile, x, cursor, w, fontPath, fontSize, padding, headerH);
-
-                case PanelSection.Separator:
-                    FillRect(x + padding, cursor, w - padding * 2f, 1f, SeparatorColor);
-                    return cursor + padding;
-
-                case PanelSection.Spacer spacer:
-                    return cursor + (spacer.Gap == PanelGap.Full ? padding : padding / 2f);
-
-                case PanelSection.Slot slot:
-                    return RenderProfileSlot(slot.Row.Label, EquipmentActions.GetAssignedDevice(pd, slot.Row.Slot), slot.Row.Slot,
-                        x, cursor, w, itemH, dpiScale, fontPath, fontSize, padding, arrowW, appState, pd, emojiFontPath);
-
-                case PanelSection.MountTelemetry:
-                    return RenderMountTelemetryIfAny(appState, pd.Mount, liveSessionState,
-                        x, cursor, w, itemH, dpiScale, fontPath, fontSize, padding);
-
-                case PanelSection.DeviceSettings ds:
-                    return RenderDeviceSettingsIfAny(appState, pd, ds.Device, ds.Label,
-                        x, cursor, w, itemH, dpiScale, fontPath, fontSize, padding);
-
-                case PanelSection.Site:
-                    return RenderSiteSection(pd, x, cursor, w, dpiScale, fontPath, fontSize, padding, itemH, buttonH, arrowW);
-
-                case PanelSection.GuideFocalLength:
-                    return RenderGuideFocalLengthSection(pd, x, cursor, w, dpiScale, fontPath, fontSize, padding, itemH);
-
-                case PanelSection.CameraTelemetry ct:
-                    return RenderCameraTelemetryIfAny(appState, ct.Camera,
-                        x, cursor, w, itemH, dpiScale, fontPath, fontSize, padding);
-
-                case PanelSection.OtaHeader oh:
-                    return RenderOtaHeaderSection(appState, pd, oh.Index, x, cursor, w, dpiScale, fontPath, fontSize, padding, itemH);
-
-                case PanelSection.OtaProps op:
-                {
-                    var otaData = pd.OTAs[op.Index];
-                    return State.EditingOtaIndex == op.Index
-                        ? RenderOtaPropertyEditors(appState, op.Index, otaData, x, cursor, w, itemH, dpiScale, fontPath, fontSize, padding, buttonH)
-                        : RenderOtaPropertiesSummary(otaData, x, cursor, w, itemH, fontPath, fontSize, padding);
-                }
-
-                case PanelSection.FilterTable ft:
-                    return RenderFilterTable(appState, ft.OtaIndex, pd, x, cursor, w, itemH, dpiScale, fontPath, fontSize, padding, buttonH);
-
-                case PanelSection.AddOta:
-                    return RenderAddOtaSection(appState, pd, x, cursor, w, y, h, dpiScale, fontPath, fontSize, padding, buttonH);
-
-                default:
-                    return cursor;
-            }
-        }
-
-        /// <summary>Profile-name header row. Returns the new cursor Y.</summary>
-        private float RenderProfileHeaderSection(Profile profile, float x, float cursor, float w,
-            string fontPath, float fontSize, float padding, float headerH)
-        {
-            DrawText(
-                $"Profile: {profile.DisplayName}".AsSpan(),
-                fontPath,
-                x + padding, cursor, w - padding * 2f, headerH,
-                fontSize * 1.1f, HeaderText, TextAlign.Near, TextAlign.Center);
-            return cursor + headerH;
-        }
-
-        /// <summary>Site latitude/longitude/elevation block (edit fields, display row, or "set site" button). Returns the new cursor Y.</summary>
-        private float RenderSiteSection(ProfileData pd, float x, float cursor, float w, float dpiScale,
-            string fontPath, float fontSize, float padding, float itemH, float buttonH, float arrowW)
-        {
-            // Site info -- clickable to edit
-            var site = EquipmentActions.GetSiteFromProfile(pd);
-            if (State.IsEditingSite)
-            {
-                // Show editable lat/lon/elevation fields
-                var fieldH = (int)(itemH * 1.2f);
-                var fieldW = (int)(w - padding * 2f);
-                var fieldX = (int)(x + padding);
-
-                var labelW_site = 50f * dpiScale;
-                var rowW_site = w - padding * 2f;
-
-                var latRow = FormRowLayout.LabeledInputRow("  Lat:", labelW_site, fieldH, padding, BaseFontSize * 0.85f, DimText);
-                RenderLayout(latRow, new RectF32(x + padding, cursor, rowW_site, fieldH), fontPath, dpiScale,
-                    drawFill: (_, r) => RenderTextInput(State.LatitudeInput, (int)r.X, (int)r.Y, (int)r.Width, (int)r.Height, fontPath, fontSize * 0.9f));
-                cursor += fieldH + 2;
-
-                var lonRow = FormRowLayout.LabeledInputRow("  Lon:", labelW_site, fieldH, padding, BaseFontSize * 0.85f, DimText);
-                RenderLayout(lonRow, new RectF32(x + padding, cursor, rowW_site, fieldH), fontPath, dpiScale,
-                    drawFill: (_, r) => RenderTextInput(State.LongitudeInput, (int)r.X, (int)r.Y, (int)r.Width, (int)r.Height, fontPath, fontSize * 0.9f));
-                cursor += fieldH + 2;
-
-                var elevRow = FormRowLayout.LabeledInputRow("  Elev:", labelW_site, fieldH, padding, BaseFontSize * 0.85f, DimText);
-                RenderLayout(elevRow, new RectF32(x + padding, cursor, rowW_site, fieldH), fontPath, dpiScale,
-                    drawFill: (_, r) => RenderTextInput(State.ElevationInput, (int)r.X, (int)r.Y, (int)r.Width, (int)r.Height, fontPath, fontSize * 0.9f));
-                cursor += fieldH + 2;
-
-                // Tie-breaker selector: which side wins on mount connect when both have a site?
-                DrawText("  Tie:".AsSpan(), fontPath, x + padding, cursor, 50f * dpiScale, itemH, fontSize * 0.85f, DimText, TextAlign.Near, TextAlign.Center);
-                var tbX = fieldX + (int)(50f * dpiScale);
-                var tbW = (fieldW - (int)(50f * dpiScale)) / 2 - 2;
-                var isMountWins = pd.SiteTieBreaker == SiteTieBreaker.Mount;
-                RenderButton("Mount", tbX, cursor, tbW, buttonH, fontPath, fontSize,
-                    isMountWins ? SlotActive : CreateButton, BodyText, "TieMount",
-                    _ => PostSignal(new UpdateProfileSignal(EquipmentActions.SetSiteTieBreaker(pd, SiteTieBreaker.Mount))));
-                RenderButton("Profile", tbX + tbW + 4, cursor, tbW, buttonH, fontPath, fontSize,
-                    !isMountWins ? SlotActive : CreateButton, BodyText, "TieProfile",
-                    _ => PostSignal(new UpdateProfileSignal(EquipmentActions.SetSiteTieBreaker(pd, SiteTieBreaker.Profile))));
-                cursor += buttonH + 2;
-
-                // Save button
-                var saveBtnW = Renderer.MeasureText("Save Site".AsSpan(), fontPath, fontSize).Width + padding * 4f;
-                RenderButton("Save Site", x + padding, cursor, saveBtnW, buttonH, fontPath, fontSize, CreateButton, BodyText, "SaveSite",
-                    _ => State.LatitudeInput.OnCommit?.Invoke(State.LatitudeInput.Text));
-                cursor += buttonH + padding;
-            }
-            else if (site.HasValue)
-            {
-                var (lat, lon, elev) = site.Value;
-                var latStr = lat >= 0 ? $"{lat:F1}\u00b0N" : $"{-lat:F1}\u00b0S";
-                var lonStr = lon >= 0 ? $"{lon:F1}\u00b0E" : $"{-lon:F1}\u00b0W";
-                var elevStr = elev.HasValue ? $", {elev.Value:F0}m" : "";
-                var siteStr = $"  Site: {latStr}, {lonStr}{elevStr}";
-
-                var siteBtnW = w - padding * 2f;
-                var siteRow = Layout.Builder.HStack(
-                    Layout.Builder.Text(siteStr, BaseFontSize * 0.9f, SiteText).Stretch(),
-                    Layout.Builder.Text("[>]", BaseFontSize * 0.85f, DimText, TextAlign.Center, TextAlign.Center).WFixed(arrowW).HStar())
-                    .WFixed(siteBtnW).HFixed(itemH)
-                    .Bg(SlotNormal)
-                    .Clickable(new HitResult.ButtonHit("EditSite"), _ => PostSignal(new EditSiteSignal()));
-                RenderLayout(siteRow, new RectF32(x + padding, cursor, siteBtnW, itemH), fontPath, dpiScale);
-                cursor += itemH;
-            }
-            else
-            {
-                // No site configured -- show "Set Site" button
-                var setSiteBtnW = Renderer.MeasureText("Set Site".AsSpan(), fontPath, fontSize).Width + padding * 4f;
-                var setSiteLeaf = Layout.Builder.Text("Set Site", BaseFontSize, BodyText, TextAlign.Center, TextAlign.Center)
-                    .WFixed(setSiteBtnW).HFixed(buttonH)
-                    .Bg(CreateButton)
-                    .Clickable(new HitResult.ButtonHit("EditSite"), _ => PostSignal(new EditSiteSignal()));
-                RenderLayout(setSiteLeaf, new RectF32(x + padding, cursor, setSiteBtnW, buttonH), fontPath, dpiScale);
-                cursor += buttonH;
-            }
-            return cursor;
-        }
-
-        /// <summary>Guide-scope focal-length input row. Returns the new cursor Y.</summary>
-        private float RenderGuideFocalLengthSection(ProfileData pd, float x, float cursor, float w, float dpiScale,
-            string fontPath, float fontSize, float padding, float itemH)
-        {
-            var labelW = w * 0.35f;
-            var fieldH = (int)(itemH * 0.9f);
-
-            // Initialize from profile if not already set
-            if (State.GuiderFocalLengthInput.Text.Length == 0 && pd.GuiderFocalLength is > 0)
-            {
-                State.GuiderFocalLengthInput.Text = pd.GuiderFocalLength.Value.ToString();
-                State.GuiderFocalLengthInput.CursorPos = State.GuiderFocalLengthInput.Text.Length;
-            }
-
-            var guideRow = FormRowLayout.LabeledInputRow("Guide FL (mm):", labelW, fieldH, padding, BaseFontSize * 0.85f, DimText);
-            RenderLayout(guideRow, new RectF32(x + padding, cursor, w - padding * 2f, fieldH), fontPath, dpiScale,
-                drawFill: (_, r) => RenderTextInput(State.GuiderFocalLengthInput, (int)r.X, (int)r.Y, (int)r.Width, (int)r.Height, fontPath, fontSize * 0.9f));
-            cursor += fieldH + 2;
-            return cursor;
-        }
-
-        /// <summary>OTA section header with the Remove / Edit buttons. Returns the new cursor Y.</summary>
-        private float RenderOtaHeaderSection(GuiAppState appState, ProfileData pd, int index, float x, float cursor, float w,
-            float dpiScale, string fontPath, float fontSize, float padding, float itemH)
-        {
-            var ota = pd.OTAs[index];
-
-            // OTA header with [Remove] + [Edit] buttons
-            var isEditingOta = State.EditingOtaIndex == index;
-            var editBtnW = 50f * dpiScale;
-            var removeBtnW = 74f * dpiScale;
-            var btnGap = 4f * dpiScale;
-            var capturedI = index;
-
-            var hub = appState.DeviceHub;
-            bool OtaDeviceConnected(Uri? u) =>
-                u is not null && u != NoneDevice.Instance.DeviceUri && hub?.IsConnected(u) == true;
-            var otaHasConnectedDevice = OtaDeviceConnected(ota.Camera) || OtaDeviceConnected(ota.Focuser)
-                || OtaDeviceConnected(ota.FilterWheel) || OtaDeviceConnected(ota.Cover);
-
-            // Title leaf (left, takes remaining width after the two buttons)
-            var titleLeaf = Layout.Builder.Text($"Telescope #{index}: {ota.Name}", BaseFontSize, HeaderText).Stretch();
-
-            // [Remove] button leaf -- disabled grey when a device is connected; armed/normal otherwise
-            Layout.Node removeLeaf;
-            if (otaHasConnectedDevice)
-            {
-                removeLeaf = Layout.Builder.Text("Remove", BaseFontSize * 0.85f, DimmedText, TextAlign.Center, TextAlign.Center)
-                    .WFixed(removeBtnW).HStar()
-                    .Bg(SegmentDisabled);
-                    // No Hit, no OnClick -- disabled
-            }
-            else
-            {
-                var armed = State.PendingRemoveOtaIndex == index;
-                removeLeaf = Layout.Builder.Text(armed ? "Confirm?" : "Remove", BaseFontSize * 0.85f, BodyText, TextAlign.Center, TextAlign.Center)
-                    .WFixed(removeBtnW).HStar()
-                    .Bg(armed ? ConfirmDangerBg : RemoveButtonBg)
-                    .Clickable(new HitResult.ButtonHit($"RemoveOta{index}"), _ =>
-                    {
-                        if (State.PendingRemoveOtaIndex == capturedI)
-                        {
-                            State.PendingRemoveOtaIndex = -1;
-                            if (appState.ActiveProfile is { Data: { } removeData })
-                            {
-                                PostSignal(new UpdateProfileSignal(EquipmentActions.RemoveOTA(removeData, capturedI)));
-                            }
-                        }
-                        else
-                        {
-                            State.PendingRemoveOtaIndex = capturedI;
-                            appState.NeedsRedraw = true;
-                        }
-                    });
-            }
-
-            // Gap leaf between Remove and Edit
-            var gapLeaf = Layout.Builder.Spacer().WFixed(btnGap).HStar();
-
-            // [Edit]/[Save] toggle button leaf
-            var editLabel = isEditingOta ? "Save" : "Edit";
-            var editLeaf = Layout.Builder.Text(editLabel, BaseFontSize * 0.85f, BodyText, TextAlign.Center, TextAlign.Center)
-                .WFixed(editBtnW).HStar()
-                .Bg(EditButtonBg)
-                .Clickable(new HitResult.ButtonHit($"EditOta{index}"), _ =>
-                {
-                    if (isEditingOta)
-                    {
-                        // Commit OTA property edits
-                        if (appState.ActiveProfile is { } prof && prof.Data is { } editData)
-                        {
-                            var newName = State.OtaNameInput.Text is { Length: > 0 } n ? n : null;
-                            int? newFl = int.TryParse(State.FocalLengthInput.Text, out var fl) && fl > 0 ? fl : null;
-                            int? newAp = int.TryParse(State.ApertureInput.Text, out var ap) ? ap : null;
-                            var newData = EquipmentActions.UpdateOTA(editData, capturedI,
-                                name: newName, focalLength: newFl, aperture: newAp);
-                            PostSignal(new UpdateProfileSignal(newData));
-                        }
-                        State.StopEditingOta();
-                    }
-                    else
-                    {
-                        State.BeginEditingOta(capturedI, ota);
-                    }
-                });
-
-            // Build the full header row as a horizontal Stack
-            var headerRow = Layout.Builder.HStack(titleLeaf, removeLeaf, gapLeaf, editLeaf)
-                .WFixed(w).HFixed(itemH)
-                .Bg(OtaHeaderBg);
-            RenderLayout(headerRow, new RectF32(x, cursor, w, itemH), fontPath, dpiScale);
-            return cursor + itemH;
-        }
-
-        /// <summary>Add-OTA action row (drawn only if it fits before the panel bottom). Returns the new cursor Y.</summary>
-        private float RenderAddOtaSection(GuiAppState appState, ProfileData pd, float x, float cursor, float w, float y, float h,
-            float dpiScale, string fontPath, float fontSize, float padding, float buttonH)
-        {
-            // [+ Add OTA] button. "Connect All" used to share this row but now lives in the app
-            // chrome (top bar) so it is reachable from every tab — see VkGuiRenderer.RenderStatusBar
-            // + EquipmentActions.ComputeConnectAllStatus.
-            var addOtaBtnY = cursor;
-            var addOtaBtnW = 120f * dpiScale;
-            if (addOtaBtnY + buttonH < y + h - padding)
-            {
-                RenderButton("+ Add OTA", x + padding, addOtaBtnY, addOtaBtnW, buttonH, fontPath, fontSize, CreateButton, BodyText, "AddOta",
-                    _ => PostSignal(new AddOtaSignal()));
-            }
-            return cursor;
-        }
-
-        /// <summary>
-        /// Renders a single profile slot row. Returns the new cursor Y.
-        /// </summary>
-        private float RenderProfileSlot(
-            string label,
-            Uri? deviceUri,
-            AssignTarget slot,
-            float x, float y, float w, float itemH,
-            float dpiScale, string fontPath,
-            float fontSize, float padding, float arrowW,
-            GuiAppState? appState = null,
-            ProfileData? profileData = null,
-            string? emojiFontPath = null)
-        {
-            var isActive = State.ActiveAssignment == slot;
-
-            // Device name -- pre-truncate to the name column width (the engine's Text leaf doesn't clip).
-            // Name column = (1 - LabelShare) of the space left after the lead pad + indicator, matching
-            // EquipmentPanelLayout.SlotRow's [pad | label .35 | name * | indicator] split.
-            var nameW = (1f - EquipmentPanelLayout.LabelShare) * (w - padding - arrowW);
-            var deviceLabel = EquipmentActions.DeviceLabel(deviceUri, registry: null);
-            var truncated = TruncateToWidth(deviceLabel, fontPath, fontSize, nameW);
-
-            // Right-edge indicator. When a device is assigned and we have access to the live
-            // hub + discovery snapshot, draw a coloured square via FillRect (font-independent)
-            // so the user sees per-slot reachability without leaving the profile panel.
-            // Active assignment-mode highlight always wins and shows the original [>] arrow.
-            EquipmentActions.DeviceReachability? slotReach = null;
-            if (!isActive && deviceUri is not null && deviceUri != NoneDevice.Instance.DeviceUri
-                && profileData is { } pd && appState is { })
-            {
-                var reach0 = EquipmentActions.GetReachability(pd, appState.DeviceHub,
-                    State.DiscoveredDevices, deviceUri);
-                if (reach0 != EquipmentActions.DeviceReachability.NotAssigned)
-                {
-                    slotReach = reach0;
-                }
-            }
-
-            // Build the row through the SHARED EquipmentPanelLayout.SlotRow so the live GPU panel and the
-            // (eventual) TUI panel render the exact same structure -- one path, guarded by the bridge's tests.
-            // SlotRow works in design units, so we render it at the real DpiScale; its indicator is a Fill
-            // slot we paint below (reachability dot, or the [>] arrow when unassigned / in assignment mode).
-            var capturedSlot = slot;
-            var capturedAppState = appState;
-            var arrowColor = isActive ? AccentInstruct : DimText;
-            var isAssigned = deviceUri is not null && deviceUri != NoneDevice.Instance.DeviceUri;
-
-            var row = EquipmentPanelLayout.SlotRow(
-                new DeviceSlotRow(label, truncated, isAssigned, slot),
-                EquipmentPanelStyle.Default,
-                activeSlot: isActive ? slot : null,
-                onSlotClick: _ => _ =>
-                {
-                    var nowActive = State.ActiveAssignment != capturedSlot;
-                    State.ActiveAssignment = nowActive ? capturedSlot : null;
-                    if (nowActive && capturedAppState is not null) ScrollActiveSlotDeviceIntoView(capturedAppState);
-                    if (capturedAppState is not null) capturedAppState.NeedsRedraw = true;
-                });
-
-            RenderLayout(row, new RectF32(x, y, w, itemH), fontPath, dpiScale,
-                drawFill: (_, r) =>
-                {
-                    if (slotReach is { } reach)
-                    {
-                        var dotColor = reach switch
-                        {
-                            EquipmentActions.DeviceReachability.Connected    => ReachConnected,
-                            EquipmentActions.DeviceReachability.Disconnected => ReachDisconnected,
-                            EquipmentActions.DeviceReachability.Offline      => ReachOffline,
-                            _                                                => DimText
-                        };
-                        var dotSize = MathF.Min(r.Height * 0.45f, arrowW * 0.55f);
-                        var dotX = r.X + (r.Width - padding / 2f - dotSize) * 0.5f;
-                        var dotY = r.Y + (r.Height - dotSize) * 0.5f;
-                        FillRect(dotX, dotY, dotSize, dotSize, dotColor);
-                    }
-                    else
-                    {
-                        DrawText(">".AsSpan(), fontPath, r.X, r.Y, r.Width - padding / 2f, r.Height,
-                            fontSize, arrowColor, TextAlign.Center, TextAlign.Center);
-                    }
-                });
-
-            // Separator line at bottom of slot (painted on top of the row background).
-            FillRect(x, y + itemH - 1f, w, 1f, SeparatorColor);
-
-            return y + itemH;
-        }
-
-        // -----------------------------------------------------------------------
-        // Right panel: device list
-        // -----------------------------------------------------------------------
-
-        private void RenderDeviceList(
-            GuiAppState appState,
-            RectF32 rect,
-            float dpiScale, string fontPath,
-            string? emojiFontPath = null)
-        {
-            var fontSize   = BaseFontSize * dpiScale;
-            var padding    = BasePadding * dpiScale;
-            var itemH      = BaseItemHeight * dpiScale;
-            var headerH    = BaseHeaderHeight * dpiScale;
-            var badgeW     = BaseBadgeWidth * dpiScale;
-            var checkW     = BaseCheckmarkWidth * dpiScale;
-            var statusW    = BaseStatusGlyphWidth * dpiScale;
-            var connBtnW   = BaseConnectBtnWidth * dpiScale;
-            var buttonH    = BaseButtonHeight * dpiScale;
-
-            var x = rect.X;
-            var y = rect.Y;
-            var w = rect.Width;
-            var h = rect.Height;
-
-            // Header
-            DrawText(
-                "Discovered Devices".AsSpan(),
-                fontPath,
-                x + padding, y, w - padding * 2f, headerH,
-                fontSize * 1.05f, HeaderText, TextAlign.Near, TextAlign.Center);
-
-            FillRect(x + padding, y + headerH - 1f, w - padding * 2f, 1f, SeparatorColor);
-
-            var listTop    = y + headerH + padding / 2f;
-            var listBottom = y + h - buttonH - padding;  // top of [Discover] button strip
-            var listH      = listBottom - listTop;        // usable list height
-
-            // Capture the list rect so the Scroll handler knows whether the wheel is over the list.
-            _deviceListRect = new RectF32(x, listTop, w, listH);
-
-            var devices  = State.DiscoveredDevices;
-            var data     = appState.ActiveProfile?.Data;
-
-            var expectedType = State.ActiveAssignment?.ExpectedDeviceType;
-            // URI of the device currently assigned to the active slot (for highlighting)
-            var activeSlotUri = State.ActiveAssignment is { } activeSlot && data is { } slotData
-                ? EquipmentActions.GetAssignedDevice(slotData, activeSlot)
-                : null;
-
-            var totalItems = devices.Count;
-
-            // Hand the controller this frame's geometry (viewport = list rect, one atom = one row); it owns
-            // the offset + wheel/drag/thumb math, and VisibleRows() yields each visible device row's rect
-            // (scrollbar column reserved, overflow clipped) -- no hand-rolled rowY / width / break here.
-            _deviceScroll.SetExtent(_deviceListRect, itemH, totalItems, dpiScale);
-
-            foreach (var (i, rowRect) in _deviceScroll.VisibleRows())
-            {
-                var rowY = rowRect.Y;     // re-bind so the row body's existing rowY / rowW references stand
-                var rowW = rowRect.Width;
-
-                var device  = devices[i];
-                var isAssigned = data is { } assignData && EquipmentActions.IsDeviceAssigned(assignData, device.DeviceUri);
-                var isWrongType = expectedType.HasValue && device.DeviceType != expectedType.Value;
-                // Highlight the device currently in the active slot
-                var isCurrentForSlot = DeviceBase.SameDevice(device.DeviceUri, activeSlotUri);
-
-                // Row background -- alternate odd/even rows for readability so the On|Off
-                // buttons line up visually with their device. Active-slot highlight wins.
-                var baseBg = (i & 1) == 0 ? DeviceRowBg : DeviceRowBgAlt;
-                // Row background as a draw-only leaf; the badge/name/status/segment content draws on top.
-                // Row select is no longer a registered clickable -- a press on the row body falls through to
-                // the controller (tap-on-release posts AssignDeviceSignal, drag scrolls); the On|Off segments
-                // + confirm strips stay registered clickables and win first via HitTestAndDispatch.
-                var rowLeaf = Layout.Builder.Spacer()
-                    .Bg(isCurrentForSlot ? SlotActive : baseBg);
-                RenderLayout(rowLeaf, new RectF32(x, rowY, rowW, itemH), fontPath, dpiScale);
-                FillRect(x, rowY + itemH - 1f, rowW, 1f, SeparatorColor);
-
-                // Type badge
-                var badgeText = DeviceTypeBadge(device.DeviceType);
-                var textColor = isWrongType ? DimmedText : BodyText;
-
-                FillRect(x + padding, rowY + itemH * 0.15f, badgeW - padding, itemH * 0.7f, BadgeBg);
-                DrawText(
-                    badgeText,
-                    fontPath,
-                    x + padding, rowY, badgeW, itemH,
-                    fontSize * 0.8f, isWrongType ? DimmedText : HeaderText, TextAlign.Center, TextAlign.Center);
-
-                // Reserve right-edge columns: [power status][On|Off button][check assigned]
-                // Even when not assigned we keep the layout stable so name widths don't jitter.
-                var rightReserved = padding + checkW + padding + connBtnW + padding + statusW;
-
-                // Device name + source moniker. The source tells the user at a glance
-                // whether this discovered row is e.g. ASCOM vs native ZWO vs Canon vs
-                // OnStep, which otherwise is only implied by the display name (and not
-                // always -- "EQ6-R Pro" could be reached via Skywatcher-native, ASCOM,
-                // or Alpaca, all with the same friendly name).
-                var nameX = x + badgeW + padding;
-                var sourceText = device.Source;
-                var sourceW = string.IsNullOrEmpty(sourceText)
-                    ? 0f
-                    : Renderer.MeasureText(sourceText.AsSpan(), fontPath, fontSize * 0.75f).Width + padding;
-                var nameW = rowW - badgeW - padding * 2f - rightReserved - sourceW;
-                DrawText(
-                    device.DisplayName.AsSpan(),
-                    fontPath,
-                    nameX, rowY, nameW, itemH,
-                    fontSize, textColor, TextAlign.Near, TextAlign.Center);
-                if (sourceW > 0f)
-                {
-                    DrawText(
-                        sourceText.AsSpan(),
-                        fontPath,
-                        nameX + nameW, rowY, sourceW, itemH,
-                        fontSize * 0.75f, DimmedText, TextAlign.Near, TextAlign.Center);
-                }
-
-                // Right-edge columns laid out from the right margin inward.
-                var checkColX  = x + rowW - padding - checkW;
-                var btnColX    = checkColX - padding - connBtnW;
-                var statusColX = btnColX - padding - statusW;
-
-                // Assigned checkmark (still answers "is this URI wired into the profile?")
-                if (isAssigned)
-                {
-                    DrawText(
-                        "\u2713".AsSpan(),
-                        fontPath,
-                        checkColX, rowY, checkW, itemH,
-                        fontSize, AssignedGreen, TextAlign.Center, TextAlign.Center);
-                }
-
-                // Reachability indicator + connect/disconnect button render for EVERY discovered
-                // row (assigned or not). Unassigned devices that get connected appear as
-                // "orphans" -- useful for ad-hoc connect of e.g. Open-Meteo without first
-                // wiring it into a profile slot.
-                var reach = EquipmentActions.GetReachability(data, appState.DeviceHub, devices, device.DeviceUri);
-                {
-                    var connectUriForPending = EquipmentActions.FindAssignedUri(data, device.DeviceUri) ?? device.DeviceUri;
-                    var pending = State.PendingTransitions.ContainsKey(connectUriForPending);
-
-                    // Status indicator (display only -- answers "is hardware reachable right now?").
-                    // Drawn as a filled square via FillRect so it renders regardless of font coverage.
-                    var glyphColor = reach switch
-                    {
-                        EquipmentActions.DeviceReachability.Connected    => ReachConnected,
-                        EquipmentActions.DeviceReachability.Disconnected => ReachDisconnected,
-                        EquipmentActions.DeviceReachability.Offline      => ReachOffline,
-                        _                                                => DimText
-                    };
-                    var dotSize = MathF.Min(itemH * 0.45f, statusW * 0.6f);
-                    var dotX = statusColX + (statusW - dotSize) * 0.5f;
-                    var dotY = rowY + (itemH - dotSize) * 0.5f;
-                    FillRect(dotX, dotY, dotSize, dotSize, glyphColor);
-
-                    if (reach == EquipmentActions.DeviceReachability.Offline)
-                    {
-                        // Offline: render OFFLINE label in place of the segmented button (no clickables).
-                        DrawText(
-                            "OFFLINE".AsSpan(),
-                            fontPath,
-                            btnColX, rowY, connBtnW, itemH,
-                            fontSize * 0.75f, ReachOffline, TextAlign.Center, TextAlign.Center);
-                    }
-                    else
-                    {
-                        // Connect with the profile URI (preserves query params like apiKey, port, baud).
-                        // The discovered device URI strips these -- we only use it to identify the device.
-                        var connectUri = EquipmentActions.FindAssignedUri(data, device.DeviceUri) ?? device.DeviceUri;
-
-                        // Pending disconnect confirmations replace the right portion of the row.
-                        // Extend the strip well past the status/button/checkmark columns and
-                        // into the (now-redundant) device-name area so the action labels fit
-                        // comfortably -- the user already knows which device they're confirming.
-                        var stripX = nameX + nameW * 0.35f;
-                        var stripW = (x + rowW) - stripX - padding;
-
-                        if (DeviceBase.SameDevice(State.PendingForceConfirm, connectUri))
-                        {
-                            // Stage 2: force-disconnect confirmation. [Cancel] on left,
-                            // destructive [REALLY FORCE] on the right -- opposite side from
-                            // where Force Off was clicked, to defeat muscle-memory escalation.
-                            RenderForceConfirmStrip(connectUri, stripX, rowY, stripW, itemH,
-                                fontPath, dpiScale);
-                        }
-                        else if (DeviceBase.SameDevice(State.PendingDisconnectConfirm, connectUri))
-                        {
-                            // Stage 1: warm-or-force confirmation.
-                            RenderDisconnectConfirmStrip(connectUri, stripX, rowY, stripW, itemH,
-                                fontPath, dpiScale, State.PendingDisconnectSafety);
-                        }
-                        else
-                        {
-                            // Telegraph disconnect risk on the Off segment itself by checking the
-                            // latest cached telemetry. If the camera reports CoolerOn or Busy,
-                            // tint Off red so the user knows clicking it will land on the
-                            // confirmation strip rather than disconnecting cleanly.
-                            var key = connectUri.GetLeftPart(UriPartial.Path);
-                            var unsafeOff = false;
-                            if (State.CameraTelemetry.TryGetValue(key, out var buf) && buf.Latest is { } latest)
-                            {
-                                unsafeOff = latest.CoolerOn || latest.Busy;
-                            }
-                            // Segmented On|Off button (encodes current state + available transition).
-                            RenderConnectSegment(connectUri, btnColX, rowY, connBtnW, itemH,
-                                fontPath, dpiScale, reach, pending, offIsUnsafe: unsafeOff);
-                        }
-                    }
-                }
-            }
-
-            // Scrollbar: the controller draws its own track + thumb at the right edge of its viewport
-            // (no-op when the list fits).
-            _deviceScroll.DrawScrollBar(FillRect);
-
-            // [Discover] button at the bottom of the list area
-            var discoverBtnY = y + h - buttonH - padding;
-            var discoverLabel = State.IsDiscovering ? "Discovering..." : "Discover";
-            var discoverBtnW = Renderer.MeasureText("Discovering...".AsSpan(), fontPath, fontSize).Width + padding * 4f;
-            var discoverBtnX = x + padding;
-
-            RenderButton(discoverLabel, discoverBtnX, discoverBtnY, discoverBtnW, buttonH, fontPath, fontSize, CreateButton, BodyText, "Discover",
-                mods => PostSignal(new DiscoverDevicesSignal(IncludeFake: (mods & InputModifier.Shift) != 0)));
-
-            // When a Cover slot is being assigned, offer the Manual Light Panel (a hand-switched dumb
-            // panel -- not discoverable) as an explicit add next to [Discover]. Once assigned it flows
-            // through the ordinary Calibrator flat path, so the user can shoot manual flats with no rig.
-            if (State.ActiveAssignment?.ExpectedDeviceType == DeviceType.CoverCalibrator)
-            {
-                const string manualLabel = "+ Manual Light Panel";
-                var manualBtnW = Renderer.MeasureText(manualLabel.AsSpan(), fontPath, fontSize).Width + padding * 4f;
-                var manualBtnX = discoverBtnX + discoverBtnW + padding;
-                RenderButton(manualLabel, manualBtnX, discoverBtnY, manualBtnW, buttonH, fontPath, fontSize, CreateButton, BodyText, "AddManualCover",
-                    _ => PostSignal(new AssignManualCoverSignal()));
-            }
-        }
-
         /// <summary>
         /// Truncates <paramref name="text"/> with a trailing ellipsis so its rendered width
         /// fits within <paramref name="maxWidth"/>. Returns the original string when it already fits.
@@ -1077,187 +399,6 @@ namespace TianWen.UI.Abstractions
                 }
             }
             return string.Concat(text.AsSpan(0, best), ellipsis);
-        }
-
-        /// <summary>
-        /// Stage-1 cooler-off confirmation strip (camera stays connected). [Warm up &amp; Off]
-        /// (green), [Force] (amber, escalates), [Cancel] (neutral).
-        /// </summary>
-        private float RenderCoolerOffConfirmStrip(
-            Uri cameraUri, float x, float y, float w, float h, string fontPath, float dpiScale)
-        {
-            var capUri = cameraUri;
-            const float font = 0.78f;
-            // Three equal inset pills; gap is a design unit (2 du) so it scales with DPI.
-            var strip = Layout.Builder.HStack(
-                FormRowLayout.InsetPillButton("Warm up & Off", BaseFontSize * font, ConfirmWarmBg, BodyText,
-                    new HitResult.ButtonHit("WarmCoolerOff"), _ => PostSignal(new WarmAndCoolerOffSignal(capUri))),
-                FormRowLayout.InsetPillButton("Force", BaseFontSize * font, ConfirmForceBg, BodyText,
-                    new HitResult.ButtonHit("EscalateForceCoolerOff"),
-                    _ =>
-                    {
-                        State.PendingCoolerOffForceConfirm = capUri;
-                        State.PendingCoolerOffConfirm = null;
-                    }),
-                FormRowLayout.InsetPillButton("Cancel", BaseFontSize * font, ConfirmCancelBg, BodyText,
-                    new HitResult.ButtonHit("CancelCoolerOff"),
-                    _ =>
-                    {
-                        State.PendingCoolerOffConfirm = null;
-                        State.PendingCoolerOffForceConfirm = null;
-                    }))
-                .WithGap(2f);
-            RenderLayout(strip, new RectF32(x, y, w, h), fontPath, dpiScale);
-
-            return y + h;
-        }
-
-        /// <summary>
-        /// Stage-2 force-cooler-off confirmation. Same anti-double-click position swap as
-        /// the disconnect force flow: [Cancel] LEFT, [REALLY FORCE] RIGHT.
-        /// </summary>
-        private float RenderCoolerOffForceStrip(
-            Uri cameraUri, float x, float y, float w, float h, string fontPath, float dpiScale)
-        {
-            var capUri = cameraUri;
-            // [Cancel] LEFT, destructive [REALLY FORCE] RIGHT (anti-double-click position swap); two pills, gap 4 du.
-            var strip = Layout.Builder.HStack(
-                FormRowLayout.InsetPillButton("Cancel", BaseFontSize * 0.8f, ConfirmCancelBg, BodyText,
-                    new HitResult.ButtonHit("CancelForceCoolerOff"),
-                    _ =>
-                    {
-                        State.PendingCoolerOffForceConfirm = null;
-                        State.PendingCoolerOffConfirm = null;
-                    }),
-                FormRowLayout.InsetPillButton("\u26A0 REALLY FORCE COOLER OFF", BaseFontSize * 0.7f, ConfirmDangerBg, BodyText,
-                    new HitResult.ButtonHit("ConfirmForceCoolerOff"),
-                    _ =>
-                    {
-                        State.PendingCoolerOffForceConfirm = null;
-                        PostSignal(new SetCoolerOffSignal(capUri));
-                    }))
-                .WithGap(4f);
-            RenderLayout(strip, new RectF32(x, y, w, h), fontPath, dpiScale);
-
-            return y + h;
-        }
-
-        /// <summary>
-        /// Stage-1 confirmation strip shown when the user clicked Off on a cooled or busy
-        /// camera. Three buttons: [Warm &amp; Off] (left, safe-green), [Force Off] (middle, amber),
-        /// [Cancel] (right, neutral). Force Off escalates to a stage-2 confirmation rendered
-        /// at a different position to prevent muscle-memory double-clicks.
-        /// </summary>
-        private void RenderDisconnectConfirmStrip(
-            Uri deviceUri, float x, float y, float w, float h,
-            string fontPath, float dpiScale,
-            EquipmentActions.DisconnectSafety safety)
-        {
-            var safetyLabel = safety switch
-            {
-                EquipmentActions.DisconnectSafety.CoolerOn   => "Warm up & Off",
-                EquipmentActions.DisconnectSafety.Busy       => "Wait & Off",
-                EquipmentActions.DisconnectSafety.BusyAndCool=> "Wait + Warm up",
-                _                                            => "Warm up & Off"
-            };
-
-            var capUri = deviceUri;
-            const float font = 0.75f;
-            // Order: [Warm/Wait & Off] [Force Off -> escalates to stage 2] [Cancel]; three equal inset pills, gap 2 du.
-            var strip = Layout.Builder.HStack(
-                FormRowLayout.InsetPillButton(safetyLabel, BaseFontSize * font, ConfirmWarmBg, BodyText,
-                    new HitResult.ButtonHit("WarmDisconnect"), _ => PostSignal(new WarmAndDisconnectDeviceSignal(capUri))),
-                FormRowLayout.InsetPillButton("Force", BaseFontSize * font, ConfirmForceBg, BodyText,
-                    new HitResult.ButtonHit("EscalateForce"),
-                    _ =>
-                    {
-                        State.PendingForceConfirm = capUri;
-                        State.PendingDisconnectConfirm = null;
-                    }),
-                FormRowLayout.InsetPillButton("Cancel", BaseFontSize * font, ConfirmCancelBg, BodyText,
-                    new HitResult.ButtonHit("CancelDisconnect"),
-                    _ =>
-                    {
-                        State.PendingDisconnectConfirm = null;
-                        State.PendingForceConfirm = null;
-                    }))
-                .WithGap(2f);
-            RenderLayout(strip, new RectF32(x, y, w, h), fontPath, dpiScale);
-        }
-
-        /// <summary>
-        /// Stage-2 force-disconnect confirmation. Layout deliberately swaps positions so the
-        /// destructive [REALLY FORCE] button lands where [Cancel] was on the previous strip,
-        /// and vice-versa -- defeats the user's muscle memory for "click the same spot twice".
-        /// </summary>
-        private void RenderForceConfirmStrip(
-            Uri deviceUri, float x, float y, float w, float h,
-            string fontPath, float dpiScale)
-        {
-            var capUri = deviceUri;
-            // [Cancel] LEFT (where [Warm & Off] was), destructive [REALLY FORCE] RIGHT (where [Cancel] was):
-            // the reversed pairing means a double-click that started on [Force Off] (middle of stage 1) lands
-            // on the [Cancel] half, never the destructive button. Two equal inset pills, gap 4 du.
-            var strip = Layout.Builder.HStack(
-                FormRowLayout.InsetPillButton("Cancel", BaseFontSize * 0.8f, ConfirmCancelBg, BodyText,
-                    new HitResult.ButtonHit("CancelForce"),
-                    _ =>
-                    {
-                        State.PendingForceConfirm = null;
-                        State.PendingDisconnectConfirm = null;
-                    }),
-                FormRowLayout.InsetPillButton("\u26A0 REALLY FORCE OFF", BaseFontSize * 0.75f, ConfirmDangerBg, BodyText,
-                    new HitResult.ButtonHit("ConfirmForce"),
-                    _ =>
-                    {
-                        State.PendingForceConfirm = null;
-                        PostSignal(new ForceDisconnectDeviceSignal(capUri));
-                    }))
-                .WithGap(4f);
-            RenderLayout(strip, new RectF32(x, y, w, h), fontPath, dpiScale);
-        }
-
-        /// <summary>
-        /// Renders the segmented On|Off connect/disconnect button. The current state's segment
-        /// is highlighted; only the *other* segment is clickable. While a transition is in flight,
-        /// the inactive segment is shown as "..." and no clickables are registered.
-        /// </summary>
-        private void RenderConnectSegment(
-            Uri deviceUri,
-            float x, float y, float w, float h,
-            string fontPath, float dpiScale,
-            EquipmentActions.DeviceReachability reach,
-            bool pending,
-            bool offIsUnsafe = false)
-        {
-            var isConnected = reach == EquipmentActions.DeviceReachability.Connected;
-
-            var onBg  = isConnected ? SegmentActive : SegmentInactive;
-            var offBg = isConnected ? SegmentInactive : SegmentActive;
-            // Telegraph that Off would land on the warm/force confirmation strip: tint the inactive Off
-            // segment red. (When isConnected, Off is the actionable segment; a darker red keeps it clickable.)
-            if (offIsUnsafe)
-            {
-                offBg = isConnected ? ConfirmDangerBg : ConfirmForceBg;
-            }
-
-            // Ellipsis on the segment we are transitioning *to*; otherwise On / Off.
-            var onLabel  = pending && !isConnected ? "\u2026" : "On";
-            var offLabel = pending &&  isConnected ? "\u2026" : "Off";
-
-            // Two inset segment pills. Both register a hit (even the active "you are here" segment swallows
-            // its click so it does not fall through to the row's AssignDeviceSignal); while a transition is
-            // pending both are inert (null Hit, so the click falls through, matching the old early-return).
-            var capturedUri = deviceUri;
-            Action<InputModifier> onAction = _ => { if (!isConnected) PostSignal(new ConnectDeviceSignal(capturedUri)); };
-            Action<InputModifier> offAction = _ => { if (isConnected) PostSignal(new DisconnectDeviceSignal(capturedUri)); };
-            var seg = Layout.Builder.HStack(
-                FormRowLayout.InsetPillButton(onLabel, BaseFontSize * 0.85f, onBg, BodyText,
-                    pending ? null : new HitResult.ButtonHit("Connect"), pending ? null : onAction),
-                FormRowLayout.InsetPillButton(offLabel, BaseFontSize * 0.85f, offBg, BodyText,
-                    pending ? null : new HitResult.ButtonHit("Disconnect"), pending ? null : offAction))
-                .WithGap(1f);
-            RenderLayout(seg, new RectF32(x, y, w, h), fontPath, dpiScale);
         }
 
         // -----------------------------------------------------------------------
@@ -1300,902 +441,6 @@ namespace TianWen.UI.Abstractions
                 fontPath,
                 x + padding, y, w - padding * 2f, h,
                 fontSize * 0.9f, textColor, TextAlign.Near, TextAlign.Center);
-        }
-
-        // -----------------------------------------------------------------------
-        // OTA property editors
-        // -----------------------------------------------------------------------
-
-        private float RenderOtaPropertiesSummary(
-            OTAData ota,
-            float x, float cursor, float w, float itemH,
-            string fontPath, float fontSize, float padding)
-        {
-            var fRatio = ota.Aperture is > 0 ? $"f/{(double)ota.FocalLength / ota.Aperture.Value:F1}" : "";
-            var designStr = ota.OpticalDesign != OpticalDesign.Unknown ? ota.OpticalDesign.ToString() : "";
-            var parts = new List<string>(3);
-            if (ota.FocalLength > 0) parts.Add($"{ota.FocalLength}mm");
-            if (ota.Aperture is > 0) parts.Add($"\u00d8{ota.Aperture}mm");
-            if (fRatio.Length > 0) parts.Add(fRatio);
-            if (designStr.Length > 0) parts.Add(designStr);
-
-            if (parts.Count > 0)
-            {
-                var summary = string.Join("  ", parts);
-                DrawText(
-                    summary.AsSpan(),
-                    fontPath,
-                    x + padding * 2f, cursor, w - padding * 3f, itemH * 0.8f,
-                    fontSize * 0.8f, DimText, TextAlign.Near, TextAlign.Center);
-                cursor += itemH * 0.8f;
-            }
-
-            return cursor;
-        }
-
-        private float RenderOtaPropertyEditors(
-            GuiAppState appState,
-            int otaIndex, OTAData ota,
-            float x, float cursor, float w, float itemH,
-            float dpiScale, string fontPath, float fontSize, float padding, float buttonH)
-        {
-            var fieldH = (int)(itemH * 1.1f);
-            var labelW = 80f * dpiScale;
-            var rowW_ota = w - padding * 2f;
-
-            // Name
-            var nameRow = FormRowLayout.LabeledInputRow("  Name:", labelW, fieldH, padding, BaseFontSize * 0.85f, DimText);
-            RenderLayout(nameRow, new RectF32(x + padding, cursor, rowW_ota, fieldH), fontPath, dpiScale,
-                drawFill: (_, r) => RenderTextInput(State.OtaNameInput, (int)r.X, (int)r.Y, (int)r.Width, (int)r.Height, fontPath, fontSize * 0.9f));
-            cursor += fieldH + 2;
-
-            // Focal length
-            var flRow = FormRowLayout.LabeledInputRow("  FL (mm):", labelW, fieldH, padding, BaseFontSize * 0.85f, DimText);
-            RenderLayout(flRow, new RectF32(x + padding, cursor, rowW_ota, fieldH), fontPath, dpiScale,
-                drawFill: (_, r) => RenderTextInput(State.FocalLengthInput, (int)r.X, (int)r.Y, (int)r.Width, (int)r.Height, fontPath, fontSize * 0.9f));
-            cursor += fieldH + 2;
-
-            // Aperture
-            var apRow = FormRowLayout.LabeledInputRow("  Aper (mm):", labelW, fieldH, padding, BaseFontSize * 0.85f, DimText);
-            RenderLayout(apRow, new RectF32(x + padding, cursor, rowW_ota, fieldH), fontPath, dpiScale,
-                drawFill: (_, r) => RenderTextInput(State.ApertureInput, (int)r.X, (int)r.Y, (int)r.Width, (int)r.Height, fontPath, fontSize * 0.9f));
-            cursor += fieldH + 2;
-
-            // Optical design cycle button
-            var designLabel = $"  Design: {ota.OpticalDesign}";
-            var designBtnW = Renderer.MeasureText(designLabel.AsSpan(), fontPath, fontSize * 0.9f).Width + padding * 4f;
-            var capturedIdx = otaIndex;
-            RenderButton(designLabel, x + padding, cursor, designBtnW, buttonH, fontPath, fontSize * 0.9f, EditButtonBg, BodyText, $"CycleDesign{otaIndex}",
-                _ =>
-                {
-                    if (appState.ActiveProfile is { } prof && prof.Data is { } data)
-                    {
-                        var currentOta = data.OTAs[capturedIdx];
-                        var nextDesign = (OpticalDesign)(((int)currentOta.OpticalDesign + 1) % 8);
-                        var newData = EquipmentActions.UpdateOTA(data, capturedIdx, opticalDesign: nextDesign);
-                        PostSignal(new UpdateProfileSignal(newData));
-                    }
-                });
-            cursor += buttonH + padding;
-
-            return cursor;
-        }
-
-        // -----------------------------------------------------------------------
-        // Filter table
-        // -----------------------------------------------------------------------
-
-        private float RenderFilterTable(
-            GuiAppState appState,
-            int otaIndex, ProfileData pd,
-            float x, float cursor, float w, float itemH,
-            float dpiScale, string fontPath, float fontSize, float padding, float buttonH)
-        {
-            var savedFilters = EquipmentActions.GetFilterConfig(pd, otaIndex);
-            var isExpanded = State.ExpandedFilterOtaIndex == otaIndex;
-            var rowH = itemH * 0.9f;
-            var capturedOtaIdx = otaIndex;
-
-            // Toggle header -- clicking expands/collapses and loads/discards editing state
-            var headerLabel = isExpanded
-                ? $"    Filters ({savedFilters.Count}) [-]"
-                : $"    Filters ({savedFilters.Count}) [+]";
-            var toggleHeader = FormRowLayout.ToggleHeaderRow(
-                headerLabel, rowH, FilterTableBg, HeaderText, BaseFontSize * 0.85f,
-                new HitResult.ButtonHit($"ToggleFilters{otaIndex}"),
-                _ =>
-                {
-                    if (isExpanded)
-                    {
-                        State.ExpandedFilterOtaIndex = -1;
-                        State.StopEditingFilters();
-                    }
-                    else
-                    {
-                        State.ExpandedFilterOtaIndex = capturedOtaIdx;
-                        State.BeginEditingFilters(savedFilters);
-                    }
-                });
-            RenderLayout(toggleHeader, new RectF32(x + padding, cursor, w - padding * 2f, rowH), fontPath, dpiScale);
-            cursor += rowH;
-
-            if (!isExpanded || State.EditingFilters is not { } filters)
-            {
-                return cursor;
-            }
-
-            // Layout columns: [#][Name                ][- offset +]
-            var slotNumW = 20f * dpiScale;
-            var contentW = w - padding * 4f;
-            var offsetGroupW = 100f * dpiScale;  // [-] value [+] grouped
-            var nameColW = contentW - slotNumW - offsetGroupW;
-            var stepBtnW = 24f * dpiScale;
-            var offsetValueW = offsetGroupW - stepBtnW * 2f;
-            var colStartX = x + padding * 2f;
-
-            // Column headers
-            FillRect(x + padding, cursor, w - padding * 2f, rowH, FilterTableBg);
-            DrawText("#".AsSpan(), fontPath, colStartX, cursor, slotNumW, rowH, fontSize * 0.75f, DimText, TextAlign.Center, TextAlign.Center);
-            DrawText("Name".AsSpan(), fontPath, colStartX + slotNumW, cursor, nameColW, rowH, fontSize * 0.75f, DimText, TextAlign.Near, TextAlign.Center);
-            DrawText("Offset".AsSpan(), fontPath, colStartX + slotNumW + nameColW, cursor, offsetGroupW, rowH, fontSize * 0.75f, DimText, TextAlign.Center, TextAlign.Center);
-            cursor += rowH;
-
-            // Filter rows
-            for (var f = 0; f < filters.Count; f++)
-            {
-                var filter = filters[f];
-                var rowBg = f % 2 == 0 ? FilterTableBg : FilterRowAlt;
-                FillRect(x + padding, cursor, w - padding * 2f, rowH, rowBg);
-
-                // Slot number
-                DrawText(
-                    (f + 1).ToString().AsSpan(),
-                    fontPath,
-                    colStartX, cursor, slotNumW, rowH,
-                    fontSize * 0.8f, DimText, TextAlign.Center, TextAlign.Center);
-
-                // Filter name -- inline text input if custom editing, otherwise clickable to open dropdown
-                var capturedF = f;
-                var nameCellX = colStartX + slotNumW;
-                var nameCellY = cursor;
-
-                if (State.CustomFilterSlotIndex == f)
-                {
-                    RenderTextInput(State.CustomFilterNameInput, (int)nameCellX, (int)cursor, (int)nameColW, (int)rowH, fontPath, fontSize * 0.8f);
-                }
-                else
-                {
-                RenderButton(EquipmentActions.FilterDisplayName(filter), nameCellX, cursor, nameColW, rowH, fontPath, fontSize * 0.8f, rowBg, BodyText, $"FilterName{otaIndex}_{f}",
-                    _ =>
-                    {
-                        var existingCustom = capturedF < filters.Count ? filters[capturedF].CustomName : null;
-                        State.FilterNameDropdown.Open(
-                            nameCellX, nameCellY + rowH, nameColW,
-                            EquipmentActions.CommonFilterNames,
-                            (idx, name) =>
-                            {
-                                if (capturedF < filters.Count)
-                                {
-                                    filters[capturedF] = new InstalledFilter(name, filters[capturedF].Position);
-                                    State.FiltersDirty = true;
-                                }
-                            },
-                            hasCustomEntry: true,
-                            onCustom: () =>
-                            {
-                                State.CustomFilterSlotIndex = capturedF;
-                                // Preserve existing custom name if re-selecting Custom...
-                                var preservedName = capturedF < filters.Count && filters[capturedF].CustomName is { } cn ? cn : "";
-                                State.CustomFilterNameInput.Text = preservedName;
-                                State.CustomFilterNameInput.CursorPos = preservedName.Length;
-                                // Signal deferred text input activation (processed in OnPostFrame)
-                                PostSignal(new ActivateTextInputSignal(State.CustomFilterNameInput));
-                            },
-                            customEntryLabel: existingCustom is { Length: > 0 } ? $"Custom: {existingCustom}" : null);
-                    });
-                }
-
-                // Offset group: [-] value [+]
-                var offsetX = colStartX + slotNumW + nameColW;
-                var offsetStr = filter.Position >= 0 ? $"+{filter.Position}" : filter.Position.ToString();
-
-                RenderButton("-", offsetX, cursor, stepBtnW, rowH, fontPath, fontSize * 0.8f, EditButtonBg, BodyText, $"FilterOffDec{otaIndex}_{f}",
-                    _ =>
-                    {
-                        if (capturedF < filters.Count)
-                        {
-                            var cur = filters[capturedF];
-                            filters[capturedF] = new InstalledFilter(cur.Filter.Name, cur.Position - 1);
-                            State.FiltersDirty = true;
-                        }
-                    });
-
-                DrawText(
-                    offsetStr.AsSpan(),
-                    fontPath,
-                    offsetX + stepBtnW, cursor, offsetValueW, rowH,
-                    fontSize * 0.8f, BodyText, TextAlign.Center, TextAlign.Center);
-
-                RenderButton("+", offsetX + stepBtnW + offsetValueW, cursor, stepBtnW, rowH, fontPath, fontSize * 0.8f, EditButtonBg, BodyText, $"FilterOffInc{otaIndex}_{f}",
-                    _ =>
-                    {
-                        if (capturedF < filters.Count)
-                        {
-                            var cur = filters[capturedF];
-                            filters[capturedF] = new InstalledFilter(cur.Filter.Name, cur.Position + 1);
-                            State.FiltersDirty = true;
-                        }
-                    });
-
-                cursor += rowH;
-            }
-
-            // Save / Cancel buttons (only shown when dirty)
-            if (State.FiltersDirty)
-            {
-                var saveBtnW = Renderer.MeasureText("Save".AsSpan(), fontPath, fontSize * 0.85f).Width + padding * 3f;
-                var cancelBtnW = Renderer.MeasureText("Cancel".AsSpan(), fontPath, fontSize * 0.85f).Width + padding * 3f;
-
-                RenderButton("Save", x + padding * 2f, cursor, saveBtnW, buttonH * 0.85f, fontPath, fontSize * 0.85f, CreateButton, BodyText, $"SaveFilters{otaIndex}",
-                    _ =>
-                    {
-                        if (appState.ActiveProfile is { } prof && prof.Data is { } data && State.EditingFilters is { } editFilters)
-                        {
-                            var newData = EquipmentActions.SetFilterConfig(data, capturedOtaIdx, editFilters);
-                            PostSignal(new UpdateProfileSignal(newData));
-                            State.FiltersDirty = false;
-                        }
-                    });
-
-                RenderButton("Cancel", x + padding * 2f + saveBtnW + padding, cursor, cancelBtnW, buttonH * 0.85f, fontPath, fontSize * 0.85f, EditButtonBg, BodyText, $"CancelFilters{otaIndex}",
-                    _ =>
-                    {
-                        // Reload from profile
-                        State.BeginEditingFilters(savedFilters);
-                    });
-
-                cursor += buttonH * 0.85f + padding / 2f;
-            }
-
-            // Wire commit for custom filter name -- on Enter, apply the name.
-            // Always re-wire to avoid stale captures after list replacement.
-            State.CustomFilterNameInput.OnCommit = (text) =>
-            {
-                if (text is { Length: > 0 } && State.EditingFilters is { } ef
-                    && State.CustomFilterSlotIndex >= 0 && State.CustomFilterSlotIndex < ef.Count)
-                {
-                    ef[State.CustomFilterSlotIndex] = new InstalledFilter(text, ef[State.CustomFilterSlotIndex].Position);
-                    State.FiltersDirty = true;
-                }
-                State.CustomFilterSlotIndex = -1;
-                State.CustomFilterNameInput.Deactivate();
-                return Task.CompletedTask;
-            };
-            State.CustomFilterNameInput.OnCancel = () =>
-            {
-                State.CustomFilterSlotIndex = -1;
-            };
-
-            return cursor;
-        }
-
-        // -----------------------------------------------------------------------
-        // Generic device settings
-        // -----------------------------------------------------------------------
-
-        /// <summary>
-        /// Renders device settings for a device URI if the device declares any <see cref="DeviceSettingDescriptor"/>s.
-        /// Returns the updated cursor Y. No-ops if the device has no settings.
-        /// </summary>
-        /// <summary>
-        /// Renders the camera cooler control + telemetry sparkline panel for the given URI,
-        /// but only when the camera is currently connected via the device hub. Hidden otherwise.
-        /// Layout: collapsible header -> readout row -> setpoint input + buttons -> temp sparkline.
-        /// </summary>
-        private float RenderCameraTelemetryIfAny(
-            GuiAppState appState,
-            Uri? cameraUri,
-            float x, float cursor, float w, float itemH,
-            float dpiScale, string fontPath, float fontSize, float padding)
-        {
-            if (cameraUri is null || cameraUri == NoneDevice.Instance.DeviceUri) return cursor;
-            if (appState.DeviceHub is not { } hub || !hub.IsConnected(cameraUri)) return cursor;
-
-            var key = cameraUri.GetLeftPart(UriPartial.Path);
-            var rowH = itemH * 0.9f;
-            var headerKey = $"CamCool_{key}";
-
-            // Toggle header -- independent of the device-settings expand state.
-            // Stored on EquipmentTabState as a separate sub-key so it doesn't clash.
-            var paneKey = $"CoolerPane:{key}";
-            var isOpen = State.ExpandedDeviceSettingsUri == paneKey;
-            var headerLabel = isOpen ? "    Cooler Control [-]" : "    Cooler Control [+]";
-
-            var camToggle = FormRowLayout.ToggleHeaderRow(
-                headerLabel, rowH, FilterTableBg, HeaderText, BaseFontSize * 0.85f,
-                new HitResult.ButtonHit(headerKey),
-                _ => State.ExpandedDeviceSettingsUri = isOpen ? null : paneKey);
-            RenderLayout(camToggle, new RectF32(x + padding, cursor, w - padding * 2f, rowH), fontPath, dpiScale);
-            cursor += rowH;
-
-            if (!isOpen) return cursor;
-
-            // Latest sample (may be null on first frame after connect)
-            State.CameraTelemetry.TryGetValue(key, out var buffer);
-            var latest = buffer?.Latest;
-
-            // ---- Readout row: 4 fixed-share cells so labels never overflow ----
-            var readoutH = rowH;
-            var readoutBg = FilterRowAlt;
-            var rowX = x + padding;
-            var rowW = w - padding * 2f;
-            FillRect(rowX, cursor, rowW, readoutH, readoutBg);
-
-            string Cell(double? v, string suffix) => v is { } d ? $"{d:F1}{suffix}" : "--";
-            var cellW = rowW / 4f;
-            var cellPad = padding;
-            var cellFs = fontSize * 0.85f;
-            var tempStr = $"CCD: {Cell(latest?.CcdTempC, "\u00b0C")}";
-            var setStr  = $"Set: {Cell(latest?.SetpointC, "\u00b0C")}";
-            var pwrStr  = latest?.CoolerPowerPct is { } p ? $"Power: {p:F0}%" : "Power: --";
-            var stateStr = latest is null ? "\u2026"
-                : (latest.Value.CoolerOn ? "Cooler ON" : "Cooler OFF");
-
-            // Belt-and-suspenders: still ellipsize per cell in case the panel shrinks below threshold.
-            var inner = cellW - cellPad;
-            var t = TruncateToWidth(tempStr,  fontPath, cellFs, inner);
-            var s = TruncateToWidth(setStr,   fontPath, cellFs, inner);
-            var pw = TruncateToWidth(pwrStr,   fontPath, cellFs, inner);
-            var st = TruncateToWidth(stateStr, fontPath, cellFs, inner);
-            DrawText(t.AsSpan(),  fontPath, rowX + cellPad,                 cursor, inner, readoutH, cellFs, BodyText, TextAlign.Near, TextAlign.Center);
-            DrawText(s.AsSpan(),  fontPath, rowX + cellW + cellPad,         cursor, inner, readoutH, cellFs, BodyText, TextAlign.Near, TextAlign.Center);
-            DrawText(pw.AsSpan(), fontPath, rowX + cellW * 2f + cellPad,    cursor, inner, readoutH, cellFs, BodyText, TextAlign.Near, TextAlign.Center);
-            DrawText(st.AsSpan(), fontPath, rowX + cellW * 3f + cellPad,    cursor, inner, readoutH, cellFs, BodyText, TextAlign.Near, TextAlign.Center);
-            cursor += readoutH;
-
-            // ---- Controls row: setpoint input + [Cool to Setpoint] + [Cooler Off] ----
-            var controlsH = rowH;
-            FillRect(x + padding, cursor, w - padding * 2f, controlsH, FilterTableBg);
-            var labelW = 80f * dpiScale;
-            DrawText("    Setpoint:".AsSpan(), fontPath,
-                x + padding, cursor, labelW, controlsH,
-                fontSize * 0.85f, DimText, TextAlign.Near, TextAlign.Center);
-
-            var inputX = x + padding + labelW;
-            var inputW = 70f * dpiScale;
-            // Lazy-init the setpoint input for this URI; default text from the latest sample.
-            if (!State.CameraSetpointInputs.TryGetValue(key, out var setpointInput))
-            {
-                setpointInput = new TextInputState { Placeholder = "-10" };
-                if (latest?.SetpointC is { } sp) setpointInput.Text = sp.ToString("F1", System.Globalization.CultureInfo.InvariantCulture);
-                State.CameraSetpointInputs[key] = setpointInput;
-            }
-            RenderTextInput(setpointInput,
-                (int)inputX, (int)cursor,
-                (int)inputW, (int)controlsH,
-                fontPath, fontSize * 0.85f);
-
-            var btnGap = 4f * dpiScale;
-            var coolBtnX = inputX + inputW + btnGap;
-            var coolBtnW = 110f * dpiScale;
-            var capUri = cameraUri;
-            RenderButton("Cool to Setpoint", coolBtnX, cursor, coolBtnW, controlsH, fontPath, fontSize * 0.78f,
-                CreateButton, BodyText, $"CoolTo_{key}",
-                _ =>
-                {
-                    var txt = setpointInput.Text;
-                    if (double.TryParse(txt, System.Globalization.NumberStyles.Float,
-                            System.Globalization.CultureInfo.InvariantCulture, out var v))
-                    {
-                        PostSignal(new SetCoolerSetpointSignal(capUri, v));
-                    }
-                });
-
-            var offBtnX = coolBtnX + coolBtnW + btnGap;
-            var offBtnW = 80f * dpiScale;
-            // Safety: tint the Cooler Off button red when the cooler is on, and route
-            // the click to the confirmation strip instead of immediate cooler-off
-            // (same condensation/thermal-shock concern as disconnect).
-            var coolerUnsafe = latest is { } l && l.CoolerOn;
-            var offBg = coolerUnsafe ? ConfirmDangerBg : EditButtonBg;
-            RenderButton("Cooler Off", offBtnX, cursor, offBtnW, controlsH, fontPath, fontSize * 0.78f,
-                offBg, BodyText, $"CoolerOff_{key}",
-                _ =>
-                {
-                    if (coolerUnsafe)
-                    {
-                        State.PendingCoolerOffConfirm = capUri;
-                        State.PendingCoolerOffForceConfirm = null;
-                    }
-                    else
-                    {
-                        PostSignal(new SetCoolerOffSignal(capUri));
-                    }
-                });
-            cursor += controlsH;
-
-            // Confirmation strip -- appears under the controls row when the user clicked
-            // Cooler Off on a cooled camera. Two stages, mirror of the disconnect flow.
-            if (DeviceBase.SameDevice(State.PendingCoolerOffForceConfirm, cameraUri))
-            {
-                cursor = RenderCoolerOffForceStrip(cameraUri, x + padding, cursor, w - padding * 2f, controlsH, fontPath, dpiScale);
-            }
-            else if (DeviceBase.SameDevice(State.PendingCoolerOffConfirm, cameraUri))
-            {
-                cursor = RenderCoolerOffConfirmStrip(cameraUri, x + padding, cursor, w - padding * 2f, controlsH, fontPath, dpiScale);
-            }
-
-            // ---- Sparkline graph ----
-            var graphH = 60f * dpiScale;
-            var graphX = x + padding;
-            var graphW = w - padding * 2f;
-            FillRect(graphX, cursor, graphW, graphH, FilterTableBg);
-            if (buffer is not null && buffer.Count >= 2)
-            {
-                RenderTemperatureSparkline(buffer, graphX, cursor, graphW, graphH, fontPath, fontSize);
-            }
-            else
-            {
-                DrawText("(sampling -- graph appears after a few seconds)".AsSpan(), fontPath,
-                    graphX, cursor, graphW, graphH,
-                    fontSize * 0.8f, DimText, TextAlign.Center, TextAlign.Center);
-            }
-            cursor += graphH + padding * 0.5f;
-
-            return cursor;
-        }
-
-        /// <summary>
-        /// Renders the mount status panel (RA/Dec/Slewing/Tracking) for the given mount URI,
-        /// but only when the mount is currently hub-connected. Hidden otherwise.
-        /// State comes from the single canonical <see cref="LiveSessionState.MountState"/>, populated
-        /// by <c>AppSignalHandler.PollPreviewTelemetry</c> while idle (equipment / sky-map / live-session
-        /// tabs visible) and by the running session's poll otherwise. The expander gives the user a
-        /// visible answer to "is the mount tracking?" without having to leave the equipment tab.
-        /// </summary>
-        private float RenderMountTelemetryIfAny(
-            GuiAppState appState,
-            Uri? mountUri,
-            LiveSessionState? liveSessionState,
-            float x, float cursor, float w, float itemH,
-            float dpiScale, string fontPath, float fontSize, float padding)
-        {
-            if (mountUri is null || mountUri == NoneDevice.Instance.DeviceUri) return cursor;
-            if (appState.DeviceHub is not { } hub || !hub.IsConnected(mountUri)) return cursor;
-            if (liveSessionState is null) return cursor;
-
-            var key = mountUri.GetLeftPart(UriPartial.Path);
-            var rowH = itemH * 0.9f;
-            var headerKey = $"MountStatus_{key}";
-            var paneKey = $"MountStatusPane:{key}";
-            var isOpen = State.ExpandedDeviceSettingsUri == paneKey;
-            var headerLabel = isOpen ? "    Mount Status [-]" : "    Mount Status [+]";
-
-            var mountToggle = FormRowLayout.ToggleHeaderRow(
-                headerLabel, rowH, FilterTableBg, HeaderText, BaseFontSize * 0.85f,
-                new HitResult.ButtonHit(headerKey),
-                _ => State.ExpandedDeviceSettingsUri = isOpen ? null : paneKey);
-            RenderLayout(mountToggle, new RectF32(x + padding, cursor, w - padding * 2f, rowH), fontPath, dpiScale);
-            cursor += rowH;
-
-            if (!isOpen) return cursor;
-
-            // Snapshot the mount state once -- the field is published atomically via Interlocked.Exchange.
-            var ms = liveSessionState.MountState;
-
-            // ---- Status badge row: Slewing | Tracking | Idle, colour-coded.
-            var badgeRowH = rowH;
-            var rowX = x + padding;
-            var rowW = w - padding * 2f;
-            FillRect(rowX, cursor, rowW, badgeRowH, FilterRowAlt);
-
-            string statusLabel;
-            RGBAColor32 statusColor;
-            if (ms.IsSlewing)
-            {
-                statusLabel = "Slewing";
-                statusColor = new RGBAColor32(0xff, 0xc8, 0x66, 0xff); // amber
-            }
-            else if (ms.IsTracking)
-            {
-                statusLabel = "Tracking";
-                statusColor = new RGBAColor32(0x66, 0xdd, 0x66, 0xff); // green
-            }
-            else
-            {
-                statusLabel = "Idle";
-                statusColor = DimText;
-            }
-
-            var labelW = 80f * dpiScale;
-            DrawText("    Status:".AsSpan(), fontPath,
-                rowX, cursor, labelW, badgeRowH,
-                fontSize * 0.85f, DimText, TextAlign.Near, TextAlign.Center);
-            DrawText(statusLabel.AsSpan(), fontPath,
-                rowX + labelW, cursor, rowW - labelW, badgeRowH,
-                fontSize * 0.95f, statusColor, TextAlign.Near, TextAlign.Center);
-            cursor += badgeRowH;
-
-            // ---- Coordinate readout row: 3 cells (RA / Dec / HA) ----
-            var coordRowH = rowH;
-            FillRect(rowX, cursor, rowW, coordRowH, FilterTableBg);
-
-            string FormatRa(double hours)
-            {
-                if (double.IsNaN(hours)) return "--";
-                var h = (int)hours;
-                var mFloat = (hours - h) * 60.0;
-                var m = (int)mFloat;
-                var s = (mFloat - m) * 60.0;
-                return $"{h:00}h{m:00}m{s:00.0}s";
-            }
-            string FormatDec(double deg)
-            {
-                if (double.IsNaN(deg)) return "--";
-                var sign = deg < 0 ? "-" : "+";
-                deg = Math.Abs(deg);
-                var d = (int)deg;
-                var mFloat = (deg - d) * 60.0;
-                var m = (int)mFloat;
-                var s = (mFloat - m) * 60.0;
-                return $"{sign}{d:00}\u00b0{m:00}'{s:00.0}\"";
-            }
-            string FormatHa(double hours)
-            {
-                if (double.IsNaN(hours)) return "--";
-                var sign = hours < 0 ? "-" : "+";
-                hours = Math.Abs(hours);
-                var h = (int)hours;
-                var m = (int)((hours - h) * 60.0);
-                return $"{sign}{h:00}h{m:00}m";
-            }
-
-            var raStr = $"RA: {FormatRa(ms.RightAscension)}";
-            var decStr = $"Dec: {FormatDec(ms.Declination)}";
-            var haStr = $"HA: {FormatHa(ms.HourAngle)}";
-            var cellW = rowW / 3f;
-            var cellPad = padding;
-            var cellFs = fontSize * 0.85f;
-            var inner = cellW - cellPad;
-            var ra = TruncateToWidth(raStr, fontPath, cellFs, inner);
-            var dc = TruncateToWidth(decStr, fontPath, cellFs, inner);
-            var ha = TruncateToWidth(haStr, fontPath, cellFs, inner);
-            DrawText(ra.AsSpan(), fontPath, rowX + cellPad,                  cursor, inner, coordRowH, cellFs, BodyText, TextAlign.Near, TextAlign.Center);
-            DrawText(dc.AsSpan(), fontPath, rowX + cellW + cellPad,          cursor, inner, coordRowH, cellFs, BodyText, TextAlign.Near, TextAlign.Center);
-            DrawText(ha.AsSpan(), fontPath, rowX + cellW * 2f + cellPad,     cursor, inner, coordRowH, cellFs, BodyText, TextAlign.Near, TextAlign.Center);
-            cursor += coordRowH + padding * 0.5f;
-
-            return cursor;
-        }
-
-        /// <summary>
-        /// Draws a CCD-temperature line + setpoint reference line in the given rect.
-        /// Y-axis spans observed min/max with a small margin; X-axis is sample order.
-        /// </summary>
-        private void RenderTemperatureSparkline(
-            CameraTelemetryBuffer buffer,
-            float x, float y, float w, float h,
-            string fontPath, float fontSize)
-        {
-            // Collect points
-            var samples = new List<CameraTelemetrySample>(buffer.Count);
-            foreach (var s in buffer.InOrder()) samples.Add(s);
-            if (samples.Count < 2) return;
-
-            double minT = double.PositiveInfinity, maxT = double.NegativeInfinity;
-            foreach (var s in samples)
-            {
-                if (s.CcdTempC is { } t) { if (t < minT) minT = t; if (t > maxT) maxT = t; }
-                if (s.SetpointC is { } sp) { if (sp < minT) minT = sp; if (sp > maxT) maxT = sp; }
-            }
-            if (double.IsInfinity(minT) || double.IsInfinity(maxT)) return;
-            if (Math.Abs(maxT - minT) < 0.5) { var mid = (minT + maxT) * 0.5; minT = mid - 0.5; maxT = mid + 0.5; }
-
-            // Add a small vertical margin.
-            var range = maxT - minT;
-            minT -= range * 0.1; maxT += range * 0.1; range = maxT - minT;
-
-            // Inset for axis labels on the left
-            var labelW = 36f;
-            var plotX = x + labelW;
-            var plotW = w - labelW - 4f;
-            var plotY = y + 4f;
-            var plotH = h - 8f;
-
-            // Axis label (min and max temperature)
-            DrawText($"{maxT:F1}".AsSpan(), fontPath, x, plotY - 6f, labelW - 4f, 14f, fontSize * 0.7f, DimText, TextAlign.Far, TextAlign.Center);
-            DrawText($"{minT:F1}".AsSpan(), fontPath, x, plotY + plotH - 8f, labelW - 4f, 14f, fontSize * 0.7f, DimText, TextAlign.Far, TextAlign.Center);
-
-            // Setpoint reference line (dashed effect via short rects)
-            var lastSp = samples[samples.Count - 1].SetpointC;
-            if (lastSp is { } setPt)
-            {
-                var spY = plotY + (float)((maxT - setPt) / range) * plotH;
-                for (var dx = 0f; dx < plotW; dx += 6f)
-                {
-                    FillRect(plotX + dx, spY, 3f, 1f, ReachDisconnected);
-                }
-            }
-
-            // Plot CCD temp as connected line segments (one tiny rect per pair).
-            for (var i = 1; i < samples.Count; i++)
-            {
-                if (samples[i - 1].CcdTempC is not { } t0) continue;
-                if (samples[i].CcdTempC is not { } t1) continue;
-                var x0 = plotX + (i - 1) * plotW / (samples.Count - 1);
-                var x1 = plotX + i * plotW / (samples.Count - 1);
-                var y0 = plotY + (float)((maxT - t0) / range) * plotH;
-                var y1 = plotY + (float)((maxT - t1) / range) * plotH;
-                DrawLineSegment(x0, y0, x1, y1, ReachConnected);
-            }
-        }
-
-        /// <summary>
-        /// Cheap line segment via a chain of 1px FillRects -- adequate for a sparkline.
-        /// </summary>
-        private void DrawLineSegment(float x0, float y0, float x1, float y1, RGBAColor32 color)
-        {
-            var dx = x1 - x0; var dy = y1 - y0;
-            var steps = (int)MathF.Ceiling(MathF.Max(MathF.Abs(dx), MathF.Abs(dy)));
-            if (steps < 1) steps = 1;
-            for (var i = 0; i <= steps; i++)
-            {
-                var t = (float)i / steps;
-                FillRect(x0 + dx * t, y0 + dy * t, 1.5f, 1.5f, color);
-            }
-        }
-
-        private float RenderDeviceSettingsIfAny(
-            GuiAppState appState,
-            ProfileData pd,
-            Uri? deviceUri,
-            string sectionLabel,
-            float x, float cursor, float w, float itemH,
-            float dpiScale, string fontPath, float fontSize, float padding)
-        {
-            if (deviceUri is null || deviceUri == NoneDevice.Instance.DeviceUri)
-            {
-                return cursor;
-            }
-
-            var device = EquipmentActions.TryDeviceFromUri(deviceUri);
-            if (device is null || device.Settings.IsDefaultOrEmpty)
-            {
-                return cursor;
-            }
-
-            return RenderDeviceSettings(appState, pd, deviceUri, device.Settings, sectionLabel,
-                x, cursor, w, itemH, dpiScale, fontPath, fontSize, padding);
-        }
-
-        /// <summary>
-        /// Renders the expandable settings pane for a device. Iterates <see cref="DeviceSettingDescriptor"/>s
-        /// and dispatches on <see cref="DeviceSettingKind"/> to render the appropriate control.
-        /// </summary>
-        private float RenderDeviceSettings(
-            GuiAppState appState,
-            ProfileData pd,
-            Uri savedDeviceUri,
-            ImmutableArray<DeviceSettingDescriptor> settings,
-            string sectionLabel,
-            float x, float cursor, float w, float itemH,
-            float dpiScale, string fontPath, float fontSize, float padding)
-        {
-            var deviceKey = savedDeviceUri.GetLeftPart(UriPartial.Path);
-            var isExpanded = State.ExpandedDeviceSettingsUri == deviceKey;
-            var rowH = itemH * 0.9f;
-
-            // Toggle header
-            var headerLabel = isExpanded ? $"    {sectionLabel} [-]" : $"    {sectionLabel} [+]";
-            var devToggle = FormRowLayout.ToggleHeaderRow(
-                headerLabel, rowH, FilterTableBg, HeaderText, BaseFontSize * 0.85f,
-                new HitResult.ButtonHit($"Toggle_{deviceKey}"),
-                _ =>
-                {
-                    if (isExpanded)
-                    {
-                        State.StopEditingDeviceSettings();
-                    }
-                    else
-                    {
-                        State.BeginEditingDeviceSettings(savedDeviceUri);
-                    }
-                });
-            RenderLayout(devToggle, new RectF32(x + padding, cursor, w - padding * 2f, rowH), fontPath, dpiScale);
-            cursor += rowH;
-
-            if (!isExpanded || State.EditingDeviceUri is not { } editingUri)
-            {
-                return cursor;
-            }
-
-            var labelW = w * 0.45f;
-            var controlX = x + padding * 2f + labelW;
-            var controlW = w - padding * 4f - labelW;
-            var labelX = x + padding * 2f;
-            var stepBtnW = 24f * dpiScale;
-
-            // Local row renderer shared by the basic pass and the advanced sub-section.
-            var rowIndex = 0;
-            float RenderSettingRow(DeviceSettingDescriptor desc, float rowCursor)
-            {
-                var rowBg = rowIndex++ % 2 == 0 ? FilterTableBg : FilterRowAlt;
-                FillRect(x + padding, rowCursor, w - padding * 2f, rowH, rowBg);
-
-                // StringEditor uses a narrower label to give more space to the text input
-                var rowLabelW = desc.Kind == DeviceSettingKind.StringEditor ? w * 0.25f : labelW;
-                var rowControlX = x + padding * 2f + rowLabelW;
-                var rowControlW = w - padding * 4f - rowLabelW;
-                DrawText($"{desc.Label}:".AsSpan(), fontPath, labelX, rowCursor, rowLabelW, rowH, fontSize * 0.85f, DimText, TextAlign.Near, TextAlign.Center);
-
-                var capturedDesc = desc;
-                switch (desc.Kind)
-                {
-                    case DeviceSettingKind.BoolToggle:
-                    case DeviceSettingKind.EnumCycle:
-                    {
-                        var valueLabel = desc.FormatValue(editingUri);
-                        RenderButton(valueLabel, controlX, rowCursor, controlW, rowH, fontPath, fontSize * 0.85f, EditButtonBg, BodyText, $"Cycle_{desc.Key}",
-                            _ =>
-                            {
-                                State.EditingDeviceUri = capturedDesc.Increment(editingUri);
-                                State.DeviceSettingsDirty = true;
-                            });
-                        break;
-                    }
-
-                    case DeviceSettingKind.IntStepper:
-                    case DeviceSettingKind.FloatStepper:
-                    case DeviceSettingKind.PercentStepper:
-                    {
-                        // [-] button
-                        if (desc.Decrement is { } decrement)
-                        {
-                            RenderButton("-", controlX, rowCursor, stepBtnW, rowH, fontPath, fontSize * 0.85f, EditButtonBg, BodyText, $"Dec_{desc.Key}",
-                                _ =>
-                                {
-                                    State.EditingDeviceUri = decrement(editingUri);
-                                    State.DeviceSettingsDirty = true;
-                                });
-                        }
-
-                        // Value label
-                        var valueText = desc.FormatValue(editingUri);
-                        var valueW = controlW - stepBtnW * 2f;
-                        DrawText(valueText.AsSpan(), fontPath, controlX + stepBtnW, rowCursor, valueW, rowH, fontSize * 0.85f, BodyText, TextAlign.Center, TextAlign.Center);
-
-                        // [+] button
-                        RenderButton("+", controlX + controlW - stepBtnW, rowCursor, stepBtnW, rowH, fontPath, fontSize * 0.85f, EditButtonBg, BodyText, $"Inc_{desc.Key}",
-                            _ =>
-                            {
-                                State.EditingDeviceUri = capturedDesc.Increment(editingUri);
-                                State.DeviceSettingsDirty = true;
-                            });
-                        break;
-                    }
-
-                    case DeviceSettingKind.StringEditor:
-                    {
-                        var isEditing = State.EditingStringSettingKey == desc.Key;
-                        if (isEditing)
-                        {
-                            // Active text input
-                            if (desc.Placeholder is { } placeholder)
-                            {
-                                State.StringSettingInput.Placeholder = placeholder;
-                            }
-                            RenderTextInput(State.StringSettingInput, (int)rowControlX, (int)rowCursor, (int)rowControlW, (int)rowH, fontPath, fontSize * 0.85f);
-                        }
-                        else
-                        {
-                            // Display current value (masked if configured) with click-to-edit
-                            var rawValue = desc.FormatValue(editingUri);
-                            var displayValue = desc.Mask && rawValue.Length > 0
-                                ? new string('*', Math.Min(rawValue.Length, 8)) + rawValue[Math.Max(0, rawValue.Length - 4)..]
-                                : rawValue;
-                            if (displayValue.Length == 0)
-                            {
-                                displayValue = desc.Placeholder ?? "(empty)";
-                            }
-                            RenderButton(displayValue, rowControlX, rowCursor, rowControlW, rowH, fontPath, fontSize * 0.85f, EditButtonBg, BodyText, $"Edit_{desc.Key}",
-                                _ =>
-                                {
-                                    State.EditingStringSettingKey = capturedDesc.Key;
-                                    State.StringSettingInput.Activate(capturedDesc.FormatValue(editingUri));
-                                });
-                        }
-                        break;
-                    }
-                }
-
-                return rowCursor + rowH;
-            }
-
-            // Basic rows first; advanced rows are deferred to a collapsed-by-default sub-section.
-            var visibleAdvancedCount = 0;
-            for (var i = 0; i < settings.Length; i++)
-            {
-                var desc = settings[i];
-
-                // Conditional visibility
-                if (desc.IsVisible is { } isVisible && !isVisible(editingUri))
-                {
-                    continue;
-                }
-                if (desc.IsAdvanced)
-                {
-                    visibleAdvancedCount++;
-                    continue;
-                }
-
-                cursor = RenderSettingRow(desc, cursor);
-            }
-
-            // Advanced sub-section: expert knobs with sensible defaults, collapsed by default.
-            if (visibleAdvancedCount > 0)
-            {
-                var advancedExpanded = State.AdvancedDeviceSettingsExpanded;
-                var advancedLabel = advancedExpanded ? "      Advanced [-]" : "      Advanced [+]";
-                var advancedToggle = FormRowLayout.ToggleHeaderRow(
-                    advancedLabel, rowH, FilterTableBg, DimText, BaseFontSize * 0.85f,
-                    new HitResult.ButtonHit($"ToggleAdvanced_{deviceKey}"),
-                    _ => State.AdvancedDeviceSettingsExpanded = !advancedExpanded);
-                RenderLayout(advancedToggle, new RectF32(x + padding, cursor, w - padding * 2f, rowH), fontPath, dpiScale);
-                cursor += rowH;
-
-                if (advancedExpanded)
-                {
-                    for (var i = 0; i < settings.Length; i++)
-                    {
-                        var desc = settings[i];
-                        if (!desc.IsAdvanced || (desc.IsVisible is { } isVisible && !isVisible(editingUri)))
-                        {
-                            continue;
-                        }
-
-                        cursor = RenderSettingRow(desc, cursor);
-                    }
-                }
-            }
-
-            // Commit any active string editor when focus moves away
-            if (State.EditingStringSettingKey is { } activeStringKey
-                && State.StringSettingInput is { IsActive: false } stringInput)
-            {
-                State.EditingDeviceUri = DeviceSettingHelper.WithQueryParam(editingUri, activeStringKey, stringInput.Text);
-                State.DeviceSettingsDirty = true;
-                State.EditingStringSettingKey = null;
-            }
-
-            // Save / Cancel buttons (only when dirty)
-            if (State.DeviceSettingsDirty)
-            {
-                var btnW = 60f * dpiScale;
-                var gap = padding;
-                var saveBtnX = x + w - padding - btnW * 2f - gap;
-                var cancelBtnX = x + w - padding - btnW;
-
-                RenderButton("Save", saveBtnX, cursor, btnW, rowH, fontPath, fontSize * 0.85f, CreateButton, BodyText, $"SaveSettings_{deviceKey}",
-                    _ =>
-                    {
-                        if (appState.ActiveProfile is { Data: { } data } && State.EditingDeviceUri is { } newUri)
-                        {
-                            var newData = EquipmentActions.UpdateDeviceUri(data, savedDeviceUri, newUri);
-                            PostSignal(new UpdateProfileSignal(newData));
-                            State.BeginEditingDeviceSettings(newUri);
-                        }
-                    });
-
-                RenderButton("Cancel", cancelBtnX, cursor, btnW, rowH, fontPath, fontSize * 0.85f, EditButtonBg, DimText, $"CancelSettings_{deviceKey}",
-                    _ =>
-                    {
-                        State.BeginEditingDeviceSettings(savedDeviceUri);
-                    });
-
-                cursor += rowH;
-            }
-
-            return cursor;
         }
 
         // -----------------------------------------------------------------------
